move atomese generation in separate file
